Refactor db_engine_specs into package
