+++ superset/db_engine_specs/gsheets.py
@@ -1,0 +1,8 @@
+from superset.db_engine_specs.sqlite import SqliteEngineSpec
+
+
+class GSheetsEngineSpec(SqliteEngineSpec):
+    """Engine for Google spreadsheets"""
+    engine = 'gsheets'
+    inner_joins = False
+    allows_subquery = False
