+++ superset/db_engine_specs/vertica.py
@@ -1,0 +1,5 @@
+from superset.db_engine_specs.postgres import PostgresBaseEngineSpec
+
+
+class VerticaEngineSpec(PostgresBaseEngineSpec):
+    engine = 'vertica'
